requirements.txt
behave
selenium
allure-behave
webdriver-manager